added database interface

+++ model/model.go
@@ -11,46 +11,46 @@ type Delivery struct {
 }
 
 type Payment struct {
-	Transaction			string 	`json:"transaction"`
-	Request_id	 		string	`json:"request_id"`
-	Currency			string	`json:"currency"`
-	Provider			string 	`json:"provider"`
-	Amount				float64 `json:"amount"`
-	Payment_dt			int 	`json:"payment_dt"`
-	Bank		 		string 	`json:"bank"`
-	Delivery_cost		float64 `json:"delivery_cost"`
-	Goods_total		 	int 	`json:"goods_total"`
-	Custom_fee		 	float64 `json:"custom_fee"`
+	Transaction		string `json:"transaction"`
+	Request_id		string `json:"request_id"`
+	Currency		string `json:"currency"`
+	Provider		string `json:"provider"`
+	Amount			int `json:"amount"`
+	Payment_dt		int `json:"payment_dt"`
+	Bank			string `json:"bank"`
+	Delivery_cost 	int `json:"delivery_cost"`
+	Goods_total		int `json:"goods_total"`
+	Custom_fee 		int `json:"custom_fee"`
 }
 
 type Item struct {
-	Chrt_id			int 	`json:"chrt_id"`
-	Track_number	string	`json:"track_number"`
-	Price			float64 `json:"price"`
-	Rid				string	`json:"rid"`
-	Name			string	`json:"name"`
-	Sale			float64	`json:"sale"`
-	Size		 	string	`json:"size"`
-	Total_price		float64	`json:"total_price"`
-	Nm_id		 	int		`json:"nm_id"`
-	Brand		 	string	`json:"brand"`
-	Status		 	int   	`json:"brand"`
+	Chrt_id			int `json:"chrt_id"`
+	Track_number	string `json:"track_number"`
+	Price			int `json:"price"`
+	Rid				string `json:"rid"`
+	Name			string `json:"name"`
+	Sale			int `json:"sale"`
+	Size		 	string `json:"size"`
+	Total_price		int	`json:"total_price"`
+	Nm_id		 	int	`json:"nm_id"`
+	Brand		 	string `json:"brand"`
+	Status		 	int `json:"brand"`
 }
 
 type Order struct{
-	Order_uid 			string 		`json:"order_uid"`
-	Track_number 		string 		`json:"track_number"`
-	Entry 				string 		`json:"entry"`
-	Delivery 			Delivery	`json:"delivery"`
-	Payment 			Payment		`json:"payment"`
-	Items 				[]Item 		`json:"items"`
-	Locale 				string 		`json:"locale"`
-	Internal_signature 	string 		`json:"internal_signature"`
-	Customer_id 		string 		`json:"customer_id"`
-	Delivery_service 	string 		`json:"delivery_service"`
-	Shardkey 			string 		`json:"shardkey"`
-	Sm_id 				int 		`json:"sm_id"`
-	Date_created 		string 		`json:"date_created"`
+	Order_uid 			string `json:"order_uid"`
+	Track_number 		string `json:"track_number"`
+	Entry 				string `json:"entry"`
+	Delivery 			Delivery `json:"delivery"`
+	Payment 			Payment `json:"payment"`
+	Items 				[]Item `json:"items"`
+	Locale 				string `json:"locale"`
+	Internal_signature 	string `json:"internal_signature"`
+	Customer_id 		string `json:"customer_id"`
+	Delivery_service 	string `json:"delivery_service"`
+	Shardkey 			string `json:"shardkey"`
+	Sm_id 				int `json:"sm_id"`
+	Date_created 		string `json:"date_created"`
 	Oof_shard 			string 		`json:"oof_shard"`
 }
